Create customercomplaint_V1
This is the initial version of the customer complaint system. 
As of right now the program can prompt a customer to enter their information and comment/complaint about the store. 
This information is sent to a file called "custTest.txt" which should be changed to match the desired database in which the comments shall be sent to. 
Further testing and refinement should be done. 

It still needs to have a retrieval function for the store manager to be able to review complaints and then send feedback to the customer.